refactor: organize in modules
